+++ src/app/modules/admin-module/models/recipe-locale-title-dialog-data.type.ts
@@ -1,0 +1,7 @@
+import { Locale } from '../../search-module/models/locale.interface';
+
+export type RecipeLocaleTitleDialogData = {
+  title: string;
+  localization: Locale;
+  onSave: (value: Locale) => void;
+};