+++ CHANGELOG.md
@@ -10,6 +10,19 @@ Run command for (without commit):
 
 ## Versions
 
+### 1.22.4
+
+- refactor and sync components to show values from form control - `TextDataFieldComponent` and `NumberFieldComponent`
+
+### 1.22.3
+
+- create dialog for localization title in `RecipeDetailComponent`
+- add configuration to show button which reset form control value to `null` in `TextDataFieldComponent`
+
+### 1.22.2
+
+- add units into locale text for `preparation time` and `calories` in `RecipeDetailComponent`
+
 ### 1.22.1
 
 - created message for Recipe approval in component `RecipeDetailComponent`